+++ CHANGELOG.md
@@ -12,6 +12,9 @@ New `functional` preset (included in `recommended`) with the following new rules
 - [no-httpinterceptors-token](https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/functional/NO_HTTPINTERCEPTORS_TOKEN.md)
 - [no-withinterceptorsfromdi](https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/functional/NO_WITHINTERCEPTORSFROMDI.md)
 
+New related rules not in the preset:
+- [no-canload-class](https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/deprecated/NO_CANLOAD_CLASS.md)
+
 ## 1.4.0 (2026-08-17)
 
 New `dependencyInjection` preset (included in `recommended`) with the following new rules: